updated attendance form and setting localDate and localTime in attend…

+++ src/main/java/com/emp/management/model/EmployeeDesignation.java
@@ -11,14 +11,15 @@ public class EmployeeDesignation {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private int id;
+	private long id;
 	
 	private String designation;
 	
-	public int getId() {
+	
+	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 	public String getDesignation() {
